move Composer lock specific behavior and extraction logic out of generator.rs
